Add bigIncremet to migrations

+++ database/migrations/2020_03_16_205326_create_tasks_table.php
@@ -16,12 +16,12 @@ class CreateTasksTable extends Migration
         Schema::create('tasks', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('category_id');
-            $table->integer('user_id');
+            $table->bigInteger('category_id');
+            $table->bigInteger('user_id');
             $table->integer('order');
             $table->timestamps();
             $table->softDeletes();
-            
+
             $table->foreign('user_id')->references('id')->on('users');
             $table->foreign('category_id')->references('id')->on('categories');
         });
